Fix build and generated code for later versions of Java.  Tested with openjdk-11, oracle-11, and oracle-12 on Ubuntu 18.04
