+++ GUI/templates/Common/libJava/Sconstruct.tpl
@@ -82,7 +82,7 @@ Swig2 =env.SharedLibrary(target = './bin/JausAddress', source = [common_dir + '/
 
 classes = env.Java('./Build', '#.')
 
-env.Jar('./bin/%file_name%.jar', ['./Build', '#./lib/Manifest.txt'] )
+env.Jar('./bin/%file_name%.jar', classes + ['./lib/Manifest.txt'] )
 
 env.Install('#./bin', '#./lib/statemap.jar')
