+++ examples/Java/ExampleOCU_200/Sconstruct
@@ -82,7 +82,7 @@ Swig2 =env.SharedLibrary(target = './bin/JausAddress', source = [common_dir + '/
 
 classes = env.Java('./Build', '#.')
 
-env.Jar('./bin/ExampleOCU_200.jar', ['./Build', '#./lib/Manifest.txt'] )
+env.Jar('./bin/ExampleOCU_200.jar', classes + ['./lib/Manifest.txt'] )
 
 env.Install('#./bin', '#./lib/statemap.jar')
