+++ examples/Java/PingComponent_120/Sconstruct
@@ -82,7 +82,7 @@ Swig2 =env.SharedLibrary(target = './bin/JausAddress', source = [common_dir + '/
 
 classes = env.Java('./Build', '#.')
 
-env.Jar('./bin/PingComponent_120.jar', ['./Build', '#./lib/Manifest.txt'] )
+env.Jar('./bin/PingComponent_120.jar', classes + ['./lib/Manifest.txt'] )
 
 env.Install('#./bin', '#./lib/statemap.jar')
